extracts UIView category in a separate file
